# Optimize flash writes on SAMD21