In its current implementation, the `machine.Flash.WriteAt()` function on SAMD21 will copy data to the internal flash memory in 4-byte chunks -- even if the page buffer is actually 64 bytes long for all chip variants.

The existing code can be improved to get ~12x faster write speeds, which will be useful when #4844 (USB mass storage) is merged and SAMD21 support is added.

I get these results when writing 4kB of data:

|        | at 48MHz | at 8MHz |
|--------|----------|---------|
| before | 229ms    | 261ms   |
| after  | 17ms     | 23ms    |

This will also make sure that no more than 4 writes are performed to the same flash _row_ (256 bytes long). The datasheet says:

> on this flash technology, a max number of 8 consecutive write is allowed per row. Once this number is reached, a row erase is mandatory.

Based on my long-term testing (working on a device that reads and writes a lot), it seems to work as expected. Data is saved to the expected memory location and can be read after the reset. It would be nice if someone else could take a look.

Of course, this is a breaking change. If some application assumes the block size of 4, it will not work properly.

There's also SAMD51 that could benefit from similar change, but I don't have any device to test with.

References:

- Datasheet: https://ww1.microchip.com/downloads/en/DeviceDoc/SAM_D21_DA1_Family_DataSheet_DS40001882F.pdf
  - "Table 10-3. SAM D21 Flash Memory Parameters"
  - "Table 37-42. Maximum Operating Frequency" and the notice below the table
  - "22.6.4.3 NVM Write"