@deadprogram 2025-05-23
Thank you very much for this improvement @mateusznowakdev 

> Of course, this is a breaking change. If some application assumes the block size of 4, it will not work properly.

External consumers should be using the `WriteBlockSize()` so should not impact any well-behaved code.

I agree that similar change for SAMD51 would benefit. I can test it if you wanted to make this change in another PR.

In any case, thanks again for working on this. Now merging.